Argument count redesign
* argc type changed from size_t to int
* Negative argc semantics (at most)
* Removed command name as empty string check
  * We aren't even checking if the first character is a separator, so
this test is really not effective. We're better off not checking the
name anyways.
* Limited argument count to 64 (static buffer)
* Added dynamic buffer abstract data type string_t in parser.c
* Strings start with size 1 and its size is increases exponentially
* Removed argvbuf (buffer of pointers to character addresses) and added
argidx (buffer of indices), which avoids dangling pointers when argument buffer
is reallocated
* yatester_raise and yatester_assert do nothing when status =
YATESTER_OK
* Added YATESTER_BADARG for failed sscanf (to differentiate from generic
errors, which might be caught with except) -- No one really expects for
a bad argument, right?
* Changed /assert <condition> to /assert <condition> <status> in
selftest

+++ src/yatester/builtins.h
@@ -9,7 +9,7 @@
  * @brief Expect next command status code
  * @note Usage: /expect <status-code>
  */
-void yatester_builtin_expect(size_t argc, char** argv);
+void yatester_builtin_expect(int argc, char** argv);
 
 /**
  * @brief NULL-terminated array of built-in commands
